refactor(studio): move transform commit handlers to a co-located util
Addresses review on #1427: the inline `commitAnimatedTransformValue` extraction
pushed PropertyPanel.tsx to 613 lines (over the 600 limit). Move the transform
commit factory (`commitAnimatedTransformValue` + the X/Y, W/H and rotation
handlers) into `propertyPanelTransformCommit.ts` and call it from the panel.
PropertyPanel.tsx drops to 543 lines; behavior is unchanged.
